Version 2.5 release
Total refactoring and improving of the Substring class, which now
contains a lot of new useful methods; new and improved old unit tests
for Substring. Refactoring of the StringTools class - now it is removed,
and its subclasses are moved into separate namespace. New method
'ReplaceAll' for the "Find and Replace" functionality. Other minor bug
fixes and improvements in the main project and unit tests.